Reverted double laod factor to float and corrected builder tests

+++ src/main/java/org/apache/datasketches/filters/quotientfilter/QuotientFilter.java
@@ -32,11 +32,11 @@ import org.apache.datasketches.filters.common.HeapBitArray;
 
 public class QuotientFilter extends Filter {
 
-  public static final double DEFAULT_LOAD_FACTOR = 0.8;
+  public static final float DEFAULT_LOAD_FACTOR = 0.8f;
 
   int lgQ_;
   int numFingerprintBits_;
-  double loadFactor_;
+  float loadFactor_;
   int numEntries_;
   int numExpansions_;
   BitArray bitArray_;
@@ -51,7 +51,7 @@ public class QuotientFilter extends Filter {
     this(lgQ, numFingerprintBits, DEFAULT_LOAD_FACTOR);
   }
 
-  public QuotientFilter(final int lgQ, final int numFingerprintBits, final double loadFactor) {
+  public QuotientFilter(final int lgQ, final int numFingerprintBits, final float loadFactor) {
     lgQ_ = lgQ;
     numFingerprintBits_ = numFingerprintBits;
     loadFactor_ = loadFactor;
@@ -149,7 +149,7 @@ public class QuotientFilter extends Filter {
     return lgQ_;
   }
 
-  public double getLoadFactor() {
+  public float getLoadFactor() {
     return loadFactor_;
   }
